move impls to impls module (prunes some dups)
